apps/sketch2bim/backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36  # Python 3.13 compatibility - pinned to avoid conflicts
psycopg[binary]>=3.2.1
alembic==1.13.1

# Async & Jobs
# Note: Celery removed - using FastAPI BackgroundTasks instead
# Note: Redis library removed - using Upstash Redis REST API via httpx (see services/redis_service.py)

# Authentication & Security
pyjwt==2.8.0
cryptography==42.0.0
python-jose[cryptography]==3.3.0

# HTTP & API
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Payments - Razorpay (replaced Stripe for India-based payments)
razorpay==1.4.2

# File Processing
pillow>=11.0.0
opencv-python-headless>=4.10.0.84
pdf2image==1.17.0
pytesseract>=0.3.10
python-magic>=0.4.27  # Optional: for content-based file type validation

# Processing Services
# (Replicate and Groq removed - not adding value, using local OpenCV processing)

# BIM/IFC
ifcopenshell>=0.8.0
numpy>=1.24.0  # Required for ifcopenshell.geom
ezdxf>=1.0.0  # For DWG/DXF export
# pythonocc-core==7.7.2  # Optional, heavy dependency

# Utils
python-dotenv==1.0.0
pydantic==2.10.4
pydantic-settings==2.6.1
email-validator==2.2.0

# WebSocket
python-socketio==5.11.0
websockets==12.0

# Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Reporting (optional)
reportlab>=4.0.0  # For PDF report generation

# Data extraction (optional)
jsonpath-rw>=1.4.0  # For JSONPath support in dataclass extraction

# Code quality (dev dependencies)
black>=24.0.0  # Code formatter
flake8>=7.0.0  # Linter
mypy>=1.8.0  # Type checker
pre-commit>=3.6.0  # Pre-commit hooks framework

# Shared backend package (monorepo)
-e ../../packages/shared-backend  # Local editable install

# Kubernetes
kubernetes>=28.1.0

# Backup utilities (optional)
boto3>=1.34.0  # For AWS S3 backup sync
azure-storage-blob>=12.19.0  # For Azure Blob Storage backup sync

